Rename domain to entity
