+++ internal/entity/translation.go
@@ -1,6 +1,6 @@
-// Package domain defines main entities for business logic (services), data base mapping and
+// Package entity defines main entities for business logic (services), data base mapping and
 // HTTP response objects if suitable. Each logic group entities in own file.
-package domain
+package entity
 
 type Translation struct {
 	Source      string `json:"source"       example:"auto"`
