+++ internal/service/translation.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/evrone/go-clean-template/internal/domain"
+	"github.com/evrone/go-clean-template/internal/entity"
 )
 
 type TranslationService struct {
@@ -20,7 +20,7 @@ func NewTranslationService(r TranslationRepo, w TranslationWebAPI) *TranslationS
 	}
 }
 
-func (s *TranslationService) History(ctx context.Context) ([]domain.Translation, error) {
+func (s *TranslationService) History(ctx context.Context) ([]entity.Translation, error) {
 	translations, err := s.repo.GetHistory(ctx)
 	if err != nil {
 		return nil, errors.Wrap(err, "TranslationService - History - s.repo.GetHistory")
@@ -29,15 +29,15 @@ func (s *TranslationService) History(ctx context.Context) ([]domain.Translation,
 	return translations, nil
 }
 
-func (s *TranslationService) Translate(ctx context.Context, t domain.Translation) (domain.Translation, error) {
+func (s *TranslationService) Translate(ctx context.Context, t entity.Translation) (entity.Translation, error) {
 	translation, err := s.webAPI.Translate(t)
 	if err != nil {
-		return domain.Translation{}, errors.Wrap(err, "TranslationService - Translate - s.webAPI.Translate")
+		return entity.Translation{}, errors.Wrap(err, "TranslationService - Translate - s.webAPI.Translate")
 	}
 
 	err = s.repo.Store(context.Background(), translation)
 	if err != nil {
-		return domain.Translation{}, errors.Wrap(err, "TranslationService - Translate - s.repo.Store")
+		return entity.Translation{}, errors.Wrap(err, "TranslationService - Translate - s.repo.Store")
 	}
 
 	return translation, nil
